fixed interacting when Energy is depleted
Now it's possible to unload the gathered resources when back home even though the energy is depleted.  There is now another return-barrier in the PlayerController.Interact() which checks if there's no energy and if the interactable object has a tree- or foodBehaviour on it, and if the if is true then return, else continue. This allows for the StorageHandlers to pass through the return-barrier!